More preprocessing work
